fix(sled): Allow the crypto-store to be used without the state-store
