+++ crates/matrix-sdk/Cargo.toml
@@ -28,8 +28,8 @@ indexeddb_cryptostore = ["matrix-sdk-indexeddb/encryption", "encryption"]
 encryption = ["matrix-sdk-base/encryption"]
 qrcode = ["encryption", "matrix-sdk-base/qrcode"]
 # TODO merge those two sled features
-sled_state_store = ["matrix-sdk-sled"]
-sled_cryptostore = ["matrix-sdk-sled/encryption", "encryption"]
+sled_state_store = ["matrix-sdk-sled/state-store"]
+sled_cryptostore = ["matrix-sdk-sled/crypto-store", "encryption"]
 markdown = ["ruma/markdown"]
 native-tls = ["reqwest/native-tls"]
 rustls-tls = ["reqwest/rustls-tls"]
